+++ cli/src/model/reference.rs
@@ -4,7 +4,7 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "reference")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = true)]
-    pub reference_id: i32,
+    pub reference_id: i64,
     pub name: String,
     pub type_: String, // type is a reserved keyword
     pub commit_hash: Option<String>,
